latest update

+++ app/Http/Controllers/BrandController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Brand;
+use App\Models\Shoe;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rules\Exists;
 
 class BrandController extends Controller
 {
@@ -16,7 +18,7 @@ class BrandController extends Controller
     public function create()
     {
         $validator = validator(request()->all(), [
-            'brand_name' => 'required',
+            'brand_name' => ['required', 'unique:brands']
         ]);
 
         if ($validator->fails()) {
@@ -24,7 +26,7 @@ class BrandController extends Controller
         }
 
         $data = new Brand;
-        $data->name = request()->brand_name;
+        $data->brand_name = request()->brand_name;
         $data->save();
 
         return redirect('/brands/add')->with('info', 'Brand Added');
@@ -32,7 +34,11 @@ class BrandController extends Controller
 
     public function delete($id)
     {
+        $shoe = Shoe::find($id);
         $brand = Brand::find($id);
+        if ($shoe) {
+            return back()->withErrors('There is a related shoe. Cannot Delete');
+        }
         $brand->delete();
         return back()->with('info', 'Brand Deleted');
     }
